Refactor imports: consolidate IterableInApp classes into a single inApp module and update references
