+++ src/inApp/IterableInAppClasses.ts
