+++ src/inApp/index.ts
@@ -1,0 +1,3 @@
+export * from './IterableInAppClasses';
+export * from './IterableInAppManager';
+export * from './IterableInAppMessage';
